move workingPreferencesHelper.js to /helpers
