+++ src/views/Apply/Forms/Parts/JurisdictionPreferences.vue
@@ -58,7 +58,7 @@ import { APPLICATION_FORM_PARTS } from '@/helpers/constants';
 import SelectionInput from '@/components/SelectionInput/SelectionInput.vue';
 import QuestionInput from '@/components/Form/QuestionInput.vue';
 import CandidateFormsMixIn from '@/views/Apply/Forms/CandidateFormsMixIn';
-import { filteredPreferences, tidyData } from '../../WorkingPreferences/workingPreferencesHelper';
+import { filteredPreferences, tidyData } from '@/helpers/workingPreferencesHelper.js';
 
 export default {
   name: 'CandidateFormJurisdictionPreferences',
