+++ src/views/Apply/Forms/Parts/LocationPreferences.vue
@@ -27,7 +27,7 @@
           :messages="{ required: `Please enter a value for ${vacancy.locationQuestion}` }"
           required
         />
-        
+
         <div v-else-if="filteredPreferences.length">
           <QuestionInput
             v-for="(item, itemIndex) in filteredPreferences"
@@ -58,7 +58,7 @@ import { APPLICATION_FORM_PARTS } from '@/helpers/constants';
 import SelectionInput from '@/components/SelectionInput/SelectionInput.vue';
 import QuestionInput from '@/components/Form/QuestionInput.vue';
 import CandidateFormsMixIn from '@/views/Apply/Forms/CandidateFormsMixIn';
-import { filteredPreferences, tidyData } from '../../WorkingPreferences/workingPreferencesHelper';
+import { filteredPreferences, tidyData } from '@/helpers/workingPreferencesHelper.js';
 
 export default {
   name: 'CandidateFormLocationPreferences',
