+++ src/views/Apply/Forms/Views/JurisdictionPreferences.vue
@@ -21,13 +21,13 @@
         :answer-sources="vacancy"
       />
     </dl>
-  </template>  
+  </template>
 </template>
 
 <script>
 import SelectionView from '@/components/SelectionInput/View.vue';
 import QuestionInputView from '@/components/QuestionInput/QuestionInputView.vue';
-import { filteredPreferences } from '../../WorkingPreferences/workingPreferencesHelper';
+import { filteredPreferences } from '@/helpers/workingPreferencesHelper.js';
 
 export default {
   name: 'CandidateFormViewJurisdictionPreferences',
