+++ src/views/Apply/Forms/Views/LocationPreferences.vue
@@ -20,13 +20,13 @@
         :data="application.locationPreferences[item.id]"
       />
     </dl>
-  </template>  
+  </template>
 </template>
 
 <script>
 import SelectionView from '@/components/SelectionInput/View.vue';
 import QuestionInputView from '@/components/QuestionInput/QuestionInputView.vue';
-import { filteredPreferences } from '../../WorkingPreferences/workingPreferencesHelper';
+import { filteredPreferences } from '@/helpers/workingPreferencesHelper.js';
 
 export default {
   name: 'CandidateFormViewLocationPreferences',
